+++ src/test/java/de/derivo/neo2rdf/console/TurtleConversionCLIWithParametersTest.java
@@ -36,7 +36,7 @@ public class TurtleConversionCLIWithParametersTest {
                 "--includeDeletedRelationshipTypes=%s".formatted(true),
                 "--deriveClassHierarchyByLabelSubsetCheck=%s".formatted(true),
                 "--derivePropertyHierarchyByRelationshipSubsetCheck=%s".formatted(true),
-                "--reifyOnlyStatementsWithAnnotations=%s".formatted(true),
+                "--reifyOnlyRelationshipsWithProperties=%s".formatted(true),
                 "--schemaOutputPath=%s".formatted(schemaOutputPath),
                 "--outputPath=%s".formatted(outputPath),
         };