+++ build-resources/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0 (2024-04-12)
+
+- New feature: By default, each Neo4j relationship is now reified in RDF by a distinct blank node.
+  Option `--reifyOnlyRelationshipsWithProperties` can be used to specify that only relationships with properties will be
+  reified in RDF.
+
 ## 1.0.0 (2024-01-25)
 
 - Initial public release.